+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- V 1.6.1.0:
+    - Added a feature to list bibliography information via `trident list --bibliography`.
+    - Added a new Server API `/bibliography` to serve bibliography information via HTTP.
 - V 1.6.0.0:
     - Added support to write gzipped EIGENSTRAT and PLINK files with `genoconvert` and `forge`. Both commands get a new option `-z` which creates gzipped output.
 - V 1.5.7.4: